Split TxBody module out of Cardano.Api.Typed
