+++ cardano-api/cardano-api.cabal
@@ -54,6 +54,7 @@ library
                         Cardano.Api.SerialiseRaw
                         Cardano.Api.SerialiseTextEnvelope
                         Cardano.Api.StakePoolMetadata
+                        Cardano.Api.TxBody
                         Cardano.Api.TxMetadata
                         Cardano.Api.Utils
                         Cardano.Api.Value
